requirements.txt
websockets
librosa
matplotlib
pygame
python-dotenv